// src/app/shared/models/user.model.ts
export interface User {
  id?: string
  username?: string
  firstName?: string
  lastName?: string
  age?: number
  email?: string
  phoneNumber?: string
  address?: string
  gender?: string
  role?: string | number
  password?: string
  postCode?: string
  token?: string
  isUpdatedProfile?: boolean
  status?: string
}